# backend/api/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "ToDo API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://localhost:4000/"
    },
    {
      "url": "https://localhost:4000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "api health",
            "content": {
              "application/json": {
                "schema": {
                  "status": "up"
                }
              }
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get ToDo list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoList"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoList"
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Add a ToDo",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "ToDo text"
                }
              },
              "required": [
                "text"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDo"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ToDo"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{todoId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a ToDo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDo"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ToDo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDo"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ToDo"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ToDo": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "task text"
          },
          "checked": {
            "type": "boolean",
            "example": true
          }
        },
        "xml": {
          "name": "ToDo"
        }
      },
      "ToDoList": {
        "type": "array",
        "items": {
          "xml": {
            "name": "todo"
          },
          "$ref": "#/components/schemas/ToDo"
        },
        "xml": {
          "name": "ToDoList"
        }
      }
    }
  }
}